Ренейм: TimetableGetter -> TimetableFactory. Вынес в интерфейс для последующего удобного создания тестов.
